packages/python/vystak-provider-azure/pyproject.toml
[project]
name = "vystak-provider-azure"
version = "0.1.0"
description = "Vystak Azure Container Apps platform provider"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "vystak>=0.1.0",
    "azure-identity>=1.15",
    "azure-mgmt-resource>=23.0",
    "azure-mgmt-containerregistry>=10.0",
    "azure-mgmt-appcontainers>=3.0",
    "azure-mgmt-loganalytics>=13.0",
    "azure-mgmt-rdbms>=10.1.1",
    "azure-mgmt-keyvault>=10.0.0",
    "azure-mgmt-msi>=7.0.0",
    "azure-mgmt-authorization>=4.0.0",
    "azure-keyvault-secrets>=4.7.0",
    "azure-core>=1.30.0",
    "docker>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vystak_provider_azure"]

[tool.uv.sources]
vystak = { workspace = true }
